neopixel: fix server logging in MacOS

+++ neopix/serial/arduino_client.py
@@ -1,11 +1,24 @@
 #!/usr/bin/python
 
 #---------------------------------------------------------
+import platform
+
+os_name = platform.system()
+if os_name == 'Linux':
+    SYSLOG_FILE = '/dev/log'
+elif os_name == 'Darwin':
+    SYSLOG_FILE = '/var/run/syslog'
+else:
+    print('ERROR:unknow os' + os_name)
+    exit(1)
+
+#---------------------------------------------------------
+
 import logging
 import logging.handlers
 my_logger = logging.getLogger('MyClientLogger')
 my_logger.setLevel(logging.DEBUG)
-handler = logging.handlers.SysLogHandler(address = '/dev/log')
+handler = logging.handlers.SysLogHandler(address = SYSLOG_FILE)
 my_logger.addHandler(handler)
 
 def mylogger(msg):
@@ -17,28 +30,28 @@ def usage():
     mylogger('Usage: arduino_client.py <arduino_ID> <cmd> <R> <G> <B>')
 
 #---------------------------------------------------------
-import socket, os, subprocess, time
+import socket, subprocess, time, os
 
 SOCKETFILE = "/tmp/socketname"
 
 def socketServerCheck():
-    if not os.path.exists(SOCKETFILE):
-        script_path = os.path.dirname(os.path.realpath(__file__))
-        mylogger('Server down')
-        server_cmd = '/usr/bin/python ' + script_path + '/arduino_server.py'
-        mylogger('Restarting ' + server_cmd)
-        subprocess.Popen(server_cmd , stdout=None, shell=True)
-        i = 0
-        while not os.path.exists(SOCKETFILE):
-            time.sleep(1)
-            i += 1
-            mylogger('.')
-            if i > 5:
-                mylogger('Server not found, EXIT')
-                exit(1)
-
-    else:
+    if os.path.exists(SOCKETFILE):
         mylogger('Server OK')
+        return;
+
+    script_path = os.path.dirname(os.path.realpath(__file__))
+    mylogger('Server down')
+    server_cmd = '/usr/bin/python ' + script_path + '/arduino_server.py'
+    mylogger('Restarting ' + server_cmd)
+    subprocess.Popen(server_cmd , stdout=None, shell=True)
+    i = 0
+    while not os.path.exists(SOCKETFILE):
+        time.sleep(1)
+        i += 1
+        mylogger('.')
+        if i > 5:
+            mylogger('Server not found, EXIT')
+            exit(1)
 
 def socketSend(msg):
     s = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
@@ -53,8 +66,11 @@ import sys
 
 if len(sys.argv) < 6:
     if len(sys.argv) == 2 and sys.argv[1] == 'kill':
-        mylogger('kill server')
-        socketSend('kill')
+        if os.path.exists(SOCKETFILE):
+            mylogger('kill server')
+            socketSend('kill')
+        else:
+            mylogger('server is down')
         exit(0)
 
     mylogger('ERROR Not enough args')
